Fri, Aug 25, 2023, 2:56 AM +08:00

+++ readability/readability.c
@@ -20,8 +20,8 @@ int main (void)
     int sentences = count_sentences(text);
 
 
-    float L = letters / words * 100;
-    float S = sentences / words * 100;
+    float L = (float) letters / (float) words * 100;
+    float S = (float) sentences / (float) words * 100;
 
     int index = round(0.0588 * L - 0.296 * S - 15.8);
